added lock to outputs, added friend and predeclarations
pre-declare `MotorPair` and `Omni` right before the `Motor` class so that the Motor class can declare them as "friends" so that they can access private variables of the `Motor` class. Also added a lock `std::map` to check if motors are "locked" by other lambda functions or threads trying to constantly update them.